Event: Use long instead of int for timestamp

+++ Energie-Efficient-Schedulers/src/main/java/at/ac/tuwien/ldsc/group1/domain/Event.java
@@ -1,17 +1,17 @@
 package at.ac.tuwien.ldsc.group1.domain;
 
 public class Event {
-    int eventTime; // time on the event time line.
+    long eventTime; // time on the event time line.
     EventType eventType;
     Application application;
 
-    public Event(int eventTime, EventType eventType, Application application) {
+    public Event(long eventTime, EventType eventType, Application application) {
         this.eventTime = eventTime;
         this.eventType = eventType;
         this.application = application;
     }
 
-    public int getEventTime() {
+    public long getEventTime() {
         return eventTime;
     }
     public EventType getEventType() {
